// data/github/current/ui/packages/notifications-inbox/notifications-inbox-v1/components-v1/list/row/__generated__/InboxListRow_v1_fragment.graphql.ts
export type CheckConclusionState = "ACTION_REQUIRED" | "CANCELLED" | "FAILURE" | "NEUTRAL" | "SKIPPED" | "STALE" | "STARTUP_FAILURE" | "SUCCESS" | "TIMED_OUT" | "%future added value";
export type CheckStatusState = "COMPLETED" | "IN_PROGRESS" | "PENDING" | "QUEUED" | "REQUESTED" | "WAITING" | "%future added value";
export type IssueState = "CLOSED" | "OPEN" | "%future added value";
export type IssueStateReason = "COMPLETED" | "NOT_PLANNED" | "REOPENED" | "%future added value";
export type NotificationReason = "APPROVAL_REQUESTED" | "ASSIGN" | "AUTHOR" | "CI_ACTIVITY" | "COMMENT" | "INVITATION" | "MANUAL" | "MEMBER_FEATURE_REQUESTED" | "MENTION" | "READY_FOR_REVIEW" | "REVIEW_REQUESTED" | "SAVED" | "SECURITY_ADVISORY_CREDIT" | "SECURITY_ALERT" | "STATE_CHANGE" | "SUBSCRIBED" | "TEAM_MENTION" | "%future added value";
export type NotificationThreadSubscriptionState = "LIST_IGNORED" | "LIST_SUBSCRIBED" | "THREAD_SUBSCRIBED" | "THREAD_TYPE_SUBSCRIBED" | "UNSUBSCRIBED" | "%future added value";
export type PullRequestState = "CLOSED" | "MERGED" | "OPEN" | "%future added value";
import { FragmentRefs } from "relay-runtime";
export type InboxListRow_v1_fragment$data = {
  readonly id: string;
  readonly isDone: boolean;
  readonly isUnread: boolean;
  readonly lastUpdatedAt: string;
  readonly list: {
    readonly __typename: "Enterprise";
    readonly slug: string;
  } | {
    readonly __typename: "Organization";
    readonly login: string;
  } | {
    readonly __typename: "Repository";
    readonly name: string;
    readonly owner: {
      readonly login: string;
    };
  } | {
    readonly __typename: "Team";
    readonly name: string;
    readonly organization: {
      readonly login: string;
    };
    readonly slug: string;
  } | {
    readonly __typename: "User";
    readonly login: string;
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
  readonly reason: NotificationReason | null | undefined;
  readonly subject: {
    readonly __typename: "AdvisoryCredit";
    readonly advisory: {
      readonly summary: string;
    };
    readonly ghsaId: string;
  } | {
    readonly __typename: "CheckSuite";
    readonly checkSuiteResult: CheckConclusionState | null | undefined;
    readonly checkSuiteState: CheckStatusState;
    readonly databaseId: number | null | undefined;
  } | {
    readonly __typename: "Commit";
    readonly id: string;
    readonly message: string;
    readonly oid: any;
  } | {
    readonly __typename: "Discussion";
    readonly id: string;
    readonly number: number;
    readonly title: string;
  } | {
    readonly __typename: "Gist";
    readonly gistName: string;
    readonly gistTitle: string | null | undefined;
    readonly id: string;
  } | {
    readonly __typename: "Issue";
    readonly id: string;
    readonly issueState: IssueState;
    readonly number: number;
    readonly stateReason: IssueStateReason | null | undefined;
    readonly title: string;
    readonly url: string;
  } | {
    readonly __typename: "MemberFeatureRequestNotification";
    readonly body: string;
    readonly title: string;
  } | {
    readonly __typename: "PullRequest";
    readonly id: string;
    readonly number: number;
    readonly prState: PullRequestState;
    readonly title: string;
  } | {
    readonly __typename: "Release";
    readonly name: string | null | undefined;
    readonly tagName: string;
  } | {
    readonly __typename: "RepositoryAdvisory";
    readonly ghsaId: string;
    readonly title: string;
  } | {
    readonly __typename: "RepositoryDependabotAlertsThread";
    readonly id: string;
  } | {
    readonly __typename: "RepositoryInvitation";
    readonly inviter: {
      readonly login: string;
    };
    readonly repositoryInvitation: {
      readonly name: string;
      readonly owner: {
        readonly login: string;
      };
    } | null | undefined;
  } | {
    readonly __typename: "RepositoryVulnerabilityAlert";
    readonly alertNumber: number;
  } | {
    readonly __typename: "SecurityAdvisory";
    readonly ghsaId: string;
    readonly summary: string;
  } | {
    readonly __typename: "TeamDiscussion";
    readonly id: string;
    readonly number: number;
    readonly title: string;
  } | {
    readonly __typename: "WorkflowRun";
    readonly runNumber: number;
    readonly workflowTitle: string | null | undefined;
  } | {
    // This will never be '%other', but we need some
    // value in case none of the concrete values match.
    readonly __typename: "%other";
  };
  readonly subscriptionStatus: NotificationThreadSubscriptionState;
  readonly summaryItemAuthor: {
    readonly avatarUrl: string;
  } | null | undefined;
  readonly summaryItemBody: string | null | undefined;
  readonly unreadItemsCount: number;
  readonly url: string;
  readonly " $fragmentType": "InboxListRow_v1_fragment";
};